Add bundle configuration for ginger, payutc and optional settings
