+++ src/SDF/BilletterieBundle/DependencyInjection/SDFBilletterieExtension.php
@@ -22,6 +22,8 @@ class SDFBilletterieExtension extends Extension
         $configuration = new Configuration();
         $config = $this->processConfiguration($configuration, $configs);
 
+        $container->setParameter('sdf_billetterie', $config);
+
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');
     }
