+++ src/SDF/BilletterieBundle/Exception/UserNotFoundException.php
@@ -10,6 +10,6 @@ class UserNotFoundException extends NotFoundHttpException
 {
 	public function __construct($message = null, Exception $previous = null, $code = 0)
 	{
-		parent::__construct($message, $previous, array(), $code);
+		parent::__construct($message, $previous, $code);
 	}
 }
