+++ src/SDF/BilletterieBundle/Resources/config/services.yml
@@ -1,4 +1,6 @@
+parameters:
+    sdf_billetterie.utils.pdf.class: SDF\BilletterieBundle\Utils\Pdf\Pdf
+
 services:
-#    sdf_billetterie.example:
-#        class: SDF\BilletterieBundle\Example
-#        arguments: [@service_id, "plain_value", %parameter%]
+    sdf_billetterie.utils.pdf:
+        class: %sdf_billetterie.utils.pdf.class%
